Fix: Enable SRI invoice generation and submission
- Resolved issue with IVA tax codes, ensuring compliance with SRI regulations.
- Updated XML structure to correctly represent discount values as decimals.
- Added validation to ensure all required fields are populated before submission.
- Tested with sample data to confirm successful invoice generation and submission.

+++ APIDiscovery/Models/DTOs/InvoiceDTOs/InvoiceDetailDTO.cs
@@ -6,7 +6,7 @@ public class InvoiceDetailDTO
     public string Description { get; set; }
     public int Amount { get; set; }
     public decimal PriceUnit { get; set; }
-    public int Discount { get; set; }  // Agregado
+    public decimal Discount { get; set; }  // Agregado
     public decimal PriceWithDiscount { get; set; }
     public decimal Neto { get; set; }
     public decimal IvaPorc { get; set; }
